orders/src/models/ticket.ts
import mongoose from "mongoose";
import { Order, OrderStatus } from "./order";
import { updateIfCurrentPlugin } from "mongoose-update-if-current";

interface TicketAttrs {
  id: string;
  title: string;
  price: number;
}

export interface TicketDoc extends mongoose.Document {
  title: string;
  price: number;
  isReserved(): Promise<boolean>;
  version: number;
}

interface TicketModel extends mongoose.Model<TicketDoc> {
  build(attrs: TicketAttrs): TicketDoc;
  findByEventVersion(event: {
    id: string;
    version: number;
  }): Promise<TicketDoc | null>;
}

const ticketSchema = new mongoose.Schema(
  {
    title: {
      type: String,
      required: true,
    },
    price: {
      type: Number,
      required: true,
      min: 0,
    },
  },
  {
    toJSON: {
      transform(doc, ret) {
        ret.id = ret._id;
        delete ret._id;
      },
    },
  }
);

//  For Optmistic Concurrency Control. We added a versioning on ticket so we can do the processing in order
ticketSchema.set("versionKey", "version");
ticketSchema.plugin(updateIfCurrentPlugin);

// Methods!
// For Creating a Ticket. We are listening for events coming from tickets srv and storing them here.
ticketSchema.statics.build = (attrs: TicketAttrs) => {
  return new Ticket({
    _id: attrs.id,
    title: attrs.title,
    price: attrs.price,
  });
};

// For Finding the ticket and making changes in it by its ID and Version.
ticketSchema.statics.findByEventVersion = (event: {
  id: string;
  version: number;
}) => {
  return Ticket.findOne({
    _id: event.id,
    version: event.version - 1
  })
};

// Custom Version Updater Function
// ticketSchema.pre('save' , function(done){
//   // @ts-ignore
//   this.$where = {
//     version : this.get('version') - 1
//   }

//   done()
// })

ticketSchema.methods.isReserved = async function () {
  // * Run a query to look at all orders. Find an order where the ticket
  // * is the ticket we just found *and* the orders status is *not* cancelled.
  // * If we found an order from that means the ticket is reserved
  const ExistingOrder = await Order.findOne({
    ticket: this,
    status: {
      $in: [
        OrderStatus.Created,
        OrderStatus.AwaitingPayment,
        OrderStatus.Complete,
      ],
    },
  });

  return !!ExistingOrder;
};

const Ticket = mongoose.model<TicketDoc, TicketModel>("Ticket", ticketSchema);

export { Ticket };
